Update workflows
PR #56 changed the artifact name and format which broke the bleeding
edge version for `qpm-action``.

This changes the build workflow to upload two sets of run artifacts:

- `release-*` artifacts which are release-ready archived versions
- `*-qpm` artifacts which are the uncompressed files that the action
downloads.

This PR also adds a bleeding release workflow that runs on any push to
main and creates a pre-release with the resulting artifacts. This can be
more convenient than having to download a run artifact, and provides a
consistent url for the bleeding builds.